requirements.txt
absl-py==2.3.1
acres==0.5.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
dnspython==2.8.0
dotenv==0.9.9
etelemetry==0.3.1
filelock==3.20.0
Flask==2.3.3
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
looseversion==1.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
protobuf==5.29.5
prov==2.1.1
psycopg2-binary==2.9.11
puremagic==1.30
pydot==4.0.1
Pygments==2.19.2
pymongo==4.6.3
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rdflib==7.2.1
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
SQLAlchemy==2.0.44
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
threadpoolctl==3.6.0
traits==7.0.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.8
wheel==0.45.1
wrapt==2.0.0
pyjwt==2.10.1
gunicorn==21.2.0
transformers>=4.40
torch>=2.6
safetensors>=0.4.3
tokenizers>=0.22.0
python-docx==1.1.2
Faker==22.6.0
tqdm==4.66.1
pytest==9.0.0
pytest-cov==7.0.0